+++ playwright/index.tsx
@@ -3,6 +3,7 @@ import { beforeMount } from "@playwright/experimental-ct-react/hooks";
 import GlobalStyle from "../src/style/global-style";
 import CarbonProvider from "../src/components/carbon-provider/carbon-provider.component";
 import I18nProvider from "../src/components/i18n-provider/i18n-provider.component";
+import TokensWrapper from "../src/components/tokens-wrapper";
 import { noTheme, sageTheme } from "../src/style/themes";
 import enGB from "../src/locales/en-gb";
 import "../src/style/fonts.css";
@@ -27,14 +28,16 @@ const mountedTheme = (theme: string) => {
 beforeMount<HooksConfig>(async ({ App, hooksConfig }) => {
   const { theme = "sage", validationRedesignOptIn } = hooksConfig || {};
   return (
-    <CarbonProvider
-      theme={mountedTheme(theme)}
-      validationRedesignOptIn={validationRedesignOptIn}
-    >
-      <GlobalStyle />
-      <I18nProvider locale={enGB}>
-        <App />
-      </I18nProvider>
-    </CarbonProvider>
+    <TokensWrapper>
+      <CarbonProvider
+        theme={mountedTheme(theme)}
+        validationRedesignOptIn={validationRedesignOptIn}
+      >
+        <GlobalStyle />
+        <I18nProvider locale={enGB}>
+          <App />
+        </I18nProvider>
+      </CarbonProvider>
+    </TokensWrapper>
   );
 });